cmake: don't set THREADS_PTHREAD_ARG=OFF anymore
The issue was that FindThreads sometimes used try_run prior to cmake 3.10.
However, the problem of actually trying to run the executable should be
fixed in cmake 3.3, where try_run first checks the CMAKE_CROSSCOMPILING
variable. Since I can't test this right now, if we find the problem is not
resolved, we should upgrade our min cmake version to 3.10.