wattsup-libftdi: fix clang failure when -Werror is used
This was unexpected b/c no warnings were previously shown.

The ultimate failure is from '-Wdeprecated-declarations', but in fact the
original problem is that the try_compile to detect whether libftdi supports
'ftdi_tcioflush' had a warning from '-Wmain-return-type'. This warning
becomes an error when '-Werror' is used, so the try_compile fails and
energymon-wattsup-libftdi tries to use the deprecated function
'ftdi_usb_purge_buffers' instead, which then also becomes an error.